Added calculateFee and fee object to Book. Changed book constructor in Books. Added Fine class.

+++ src/Library/Fine.java
@@ -21,7 +21,7 @@ public class Fine {
         this.cost = 0;
     }
 
-    public void addCost(Integer weeks){
+    public void addCost(long weeks){
         if(this.getCost() == 0){
             this.cost += INITIALFEE;
         }
